newest version

+++ student-distrib/device.h
@@ -69,11 +69,11 @@ void init_rtc(void);
 /*change the rate of rtc freqency*/
 void set_rate(unsigned rate);
 
-uint8_t pit_interrupt();
+uint32_t pit_interrupt();
 
 void init_pit(unsigned rate);
 
-uint8_t scheduling();
+uint32_t scheduling();
 
 /*get the keyboard buffer*/
 unsigned char *getBuffer(void);
